pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "project-creator"
version = "0.1.0"

authors = [
  { name = "Oleksandra", email = "hotfish201478@gmail.com" }
]
description = "A Python CLI tool to scaffold project directories using YAML templates."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PyYAML >= 6.0"
]

[project.scripts]
scaffold = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", ".venv*", ".idea*"]